+++ libavfilter/dnn/dnn_backend_native_layer_depthwiseconv2dnative.c
@@ -17,11 +17,12 @@
  * License along with FFmpeg; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
-
+#include <stdio.h>
+#include <string.h>
 #include "libavutil/avassert.h"
-#include "dnn_backend_native_layer_depthwise.h"
+#include "dnn_backend_native_layer_depthwiseconv2dnative.h"
 
-int dnn_load_layer_depthwise(Layer *layer, AVIOContext *model_file_context, int file_size)
+int dnn_load_layer_depthwiseconv2dnative(Layer *layer, AVIOContext *model_file_context, int file_size)
 {
     DepthwiseConvParams *conv_params;
     int kernel_size;
@@ -29,7 +30,7 @@ int dnn_load_layer_depthwise(Layer *layer, AVIOContext *model_file_context, int 
     conv_params = av_malloc(sizeof(*conv_params));
     if (!conv_params)
         return 0;
-
+    printf("reaching here!\n");
     conv_params->dilation = (int32_t)avio_rl32(model_file_context);
     conv_params->padding_method = (int32_t)avio_rl32(model_file_context);
     conv_params->channel_multiplier = (int32_t)avio_rl32(model_file_context);
@@ -59,13 +60,13 @@ int dnn_load_layer_depthwise(Layer *layer, AVIOContext *model_file_context, int 
     layer->params = conv_params;
 
     layer->input_operand_indexes[0] = (int32_t)avio_rl32(model_file_context);
-    // layer->output_operand_index = (int32_t)avio_rl32(model_file_context);
-    dnn_size += 4; // change?
+    layer->output_operand_index = (int32_t)avio_rl32(model_file_context);
+    dnn_size += 8;
     return dnn_size;
 }
 
-int dnn_execute_layer_depthwise(DnnOperand *operands, const int32_t *input_operand_indexes,
-                             int32_t output_operand_index, const void *parameters)
+int dnn_execute_layer_depthwiseconv2dnative(DnnOperand *operands, const int32_t *input_operand_indexes,
+                                            int32_t output_operand_index, const void *parameters)
 {
 	float *output;
     int32_t input_operand_index = input_operand_indexes[0];
@@ -75,7 +76,6 @@ int dnn_execute_layer_depthwise(DnnOperand *operands, const int32_t *input_opera
     int in_channel = operands[input_operand_index].dims[3];
     const float *input = operands[input_operand_index].data;
     const DepthwiseConvParams *conv_params = (const DepthwiseConvParams *)parameters;
-
 
 	int pad_size = (conv_params->padding_method == VALID) ? 
 					(conv_params->kernel_size - 1) / 2 * conv_params->dilation : 0;
@@ -100,7 +100,7 @@ int dnn_execute_layer_depthwise(DnnOperand *operands, const int32_t *input_opera
     int filter_linesize = conv_params->kernel_size * in_channel;
     int src_linesize = in_width * in_channel;
 
-    float *temp_out = malloc(sizeof(float)*out_height*out_width*out_channel);
+    float *temp_out = av_malloc(sizeof(float)*out_height*out_width*out_channel);
 	int temp_out_loc = 0;
 	for (int m = 0; m<conv_params->channel_multiplier; ++m){
 	    for (int y = pad_size; y < in_height - pad_size; ++y) {
@@ -124,17 +124,18 @@ int dnn_execute_layer_depthwise(DnnOperand *operands, const int32_t *input_opera
 	   	}
 	}
 
-	int out_size_wc = out_width*out_channel;
-	int out_size_hwc = out_height*out_size_wc;
-	int out_size_nhwc = out_num*out_size_hwc;
+    //Concatenate temp_out to obtain output
+	int temp_size_wc = out_width * in_channel;
+	int temp_size_hwc = out_height * temp_size_wc;
+	int temp_size_nhwc = out_num * temp_size_hwc;
 
 	int output_ind = 0;
 	for (int n=0; n<out_num; n++){
 		for (int h=0; h<out_height; h++){
 			for (int w=0; w<out_width; w++){
-				for (int ch=0; ch<out_channel; ch++){
+				for (int ch=0; ch<in_channel; ch++){
 					for (int i=0; i<conv_params->channel_multiplier; i++){
-						int input_ind = i*out_size_nhwc + n*out_size_hwc + h*out_size_wc + w*out_channel + ch;
+						int input_ind = i*temp_size_nhwc + n*temp_size_hwc + h*temp_size_wc + w*in_channel + ch;
 						output[output_ind] = temp_out[input_ind];
 						output_ind ++;
 					}
@@ -142,7 +143,7 @@ int dnn_execute_layer_depthwise(DnnOperand *operands, const int32_t *input_opera
 			}
 		}
 	}
-	free(temp_out);
+	av_freep(&temp_out);
 	return 0;
 }
